Refactor async_connection class
